pyproject.toml
[project]
name = "fal-studio"
version = "0.1.0"
description = "fal.ai API image generation CLI"
requires-python = ">=3.11"
dependencies = [
    "fal-client>=0.5.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
fal-studio = "fal_studio.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fal_studio"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
